backend/requirements.txt
fastapi
uvicorn[standard]
openai
transformers
sentence-transformers
faiss-cpu
huggingface-hub
httpx
aiohttp
python-dotenv
fastapi[all]
aiofiles
pydantic
pydantic-settings
sqlalchemy
asyncpg
python-jose[cryptography]
jwt
passlib[bcrypt]
pytest
loguru
jupyterlab
tiktoken
langchain[google-genai]
langchain-google-genai
langchain[openai]
langchain
langchain_community
langchain_core
langchain-pinecone
langgraph
chromadb
pypdf
umap-learn
numpy 
pandas
scikit-learn
ragatouille
tavily-python
nomic[local]
langchain-nomic
protobuf==3.20.3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

